coaching_system/requirements.txt
uagents>=0.13.0
google-generativeai>=0.8.0
httpx>=0.27.0
Pillow>=10.0.0
mss>=9.0.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
